backend/python/requirements.txt
speechrecognition
deep-translator
python-dotenv
perplexityai
gtts
playsound
pyaudio
requests
